升级项目依赖，可以添加自定义的telegram bot api服务器地址

+++ TelegramSearchBot/Controller/AutoOCRController.cs
@@ -45,7 +45,7 @@ namespace TelegramSearchBot.Controller {
                                 ChatId = e.Message.Chat.Id,
                                 Content = str,
                                 MessageId = message.MessageId,
-                                UserId = botClient.BotId
+                                UserId = (long)botClient.BotId
                             });
                         }, e.Message.Chat.Id < 0);
                     }
